Refactor component code
Refactoring of components by splitting their code into small subcomponents in order to make the application more readable and scalable
