+++ src/components/Hamburger.jsx
@@ -1,8 +1,8 @@
-import { RiMenu3Fill } from "react-icons/ri";
+import { RiMenu4Line } from "react-icons/ri";
 
 function Hamburger(props) {
   return (
-    <RiMenu3Fill
+    <RiMenu4Line
       className="block md:hidden w-10 h-10 cursor-pointer stroke-0 fill-std-green"
       onClick={() => props.setMenu(true)}
     />
