+++ src/components/MenuNavigation.jsx
@@ -1,0 +1,30 @@
+import fetcher from "../services/fetcher";
+import { getData } from "../services/getData";
+import { config } from "../../config";
+
+function MenuNavigation() {
+  const navigationEndpoint = config.apiEndpoints.navigation;
+  const navigationData = getData(navigationEndpoint, fetcher);
+
+  return (
+    <nav>
+      <ul className="flex gap-12 flex-col">
+        {navigationData.map((navigationItem) => {
+          return (
+            <li
+              className="group flex items-center gap-8 pb-3 border-b-2 border-b-transparent cursor-pointer text-3xl hover:border-b-std-navy transition-all"
+              key={navigationItem.id}
+            >
+              <span className="w-14 h-14 flex items-center justify-center rounded-full bg-std-navy text-std-green">
+                {navigationItem.id}
+              </span>
+              <h1 className="text-std-navy">{navigationItem.name}</h1>
+            </li>
+          );
+        })}
+      </ul>
+    </nav>
+  );
+}
+
+export default MenuNavigation;
